+++ src/CslaGeneratorSerialization.Tests/Extensions/BinaryWriterExtensionsTests.cs
@@ -19,11 +19,11 @@ internal static class BinaryWriterExtensionsTests
 		stream.Position = 0;
 
 		var lengthBuffer = new byte[4];
-	  _ = await stream.ReadAsync(lengthBuffer.AsMemory(0, 4));
+		_ = await stream.ReadAsync(lengthBuffer.AsMemory(0, 4));
 		var length = BitConverter.ToInt32(lengthBuffer);
 
 		var content = new byte[length];
-	  _ = await stream.ReadAsync(content.AsMemory(0, length));
+		_ = await stream.ReadAsync(content.AsMemory(0, length));
 
 		Assert.That(new BigInteger(content), Is.EqualTo(number));
 	}
@@ -42,13 +42,13 @@ internal static class BinaryWriterExtensionsTests
 		using (Assert.EnterMultipleScope())
 		{
 			var lengthBuffer = new byte[4];
-		 _ = await stream.ReadAsync(lengthBuffer.AsMemory(0, 4));
+			_ = await stream.ReadAsync(lengthBuffer.AsMemory(0, 4));
 			var length = BitConverter.ToInt32(lengthBuffer);
 
 			Assert.That(length, Is.EqualTo(3));
 
 			var readBuffer = new byte[3];
-		 _ = await stream.ReadAsync(readBuffer.AsMemory(0, 3));
+			_ = await stream.ReadAsync(readBuffer.AsMemory(0, 3));
 			Assert.That(readBuffer, Is.EquivalentTo(buffer));
 		}
 	}
@@ -67,17 +67,179 @@ internal static class BinaryWriterExtensionsTests
 		using (Assert.EnterMultipleScope())
 		{
 			var lengthBuffer = new byte[4];
-		 _ = await stream.ReadAsync(lengthBuffer.AsMemory(0, 4));
+			_ = await stream.ReadAsync(lengthBuffer.AsMemory(0, 4));
 			var length = BitConverter.ToInt32(lengthBuffer);
 
 			Assert.That(length, Is.EqualTo(3));
 
 			var readBuffer = new byte[3];
-		 _ = await stream.ReadAsync(readBuffer.AsMemory(0, 3));
+			_ = await stream.ReadAsync(readBuffer.AsMemory(0, 3));
 			Assert.That(readBuffer, Has.Length.EqualTo(3));
 			Assert.That((char)readBuffer[0], Is.EqualTo(buffer[0]));
 			Assert.That((char)readBuffer[1], Is.EqualTo(buffer[1]));
 			Assert.That((char)readBuffer[2], Is.EqualTo(buffer[2]));
+		}
+	}
+
+	[Test]
+	public static async Task WriteInt16ArrayAsync()
+	{
+		short[] buffer = [1, 2, 3];
+
+		var stream = new MemoryStream();
+		using var writer = new BinaryWriter(stream);
+		writer.Write((buffer.Length, buffer));
+
+		stream.Position = 0;
+
+		using (Assert.EnterMultipleScope())
+		{
+			var lengthBuffer = new byte[4];
+			_ = await stream.ReadAsync(lengthBuffer.AsMemory(0, 4));
+			var length = BitConverter.ToInt32(lengthBuffer);
+
+			Assert.That(length, Is.EqualTo(3));
+
+			var readBuffer = new byte[6];
+			_ = await stream.ReadAsync(readBuffer.AsMemory(0, 6));
+			var newBuffer = new short[3];
+			Buffer.BlockCopy(readBuffer, 0, newBuffer, 0, 6);
+			Assert.That(newBuffer, Is.EquivalentTo(buffer));
+		}
+	}
+
+	[Test]
+	public static async Task WriteUInt16ArrayAsync()
+	{
+		ushort[] buffer = [1, 2, 3];
+
+		var stream = new MemoryStream();
+		using var writer = new BinaryWriter(stream);
+		writer.Write((buffer.Length, buffer));
+
+		stream.Position = 0;
+
+		using (Assert.EnterMultipleScope())
+		{
+			var lengthBuffer = new byte[4];
+			_ = await stream.ReadAsync(lengthBuffer.AsMemory(0, 4));
+			var length = BitConverter.ToInt32(lengthBuffer);
+
+			Assert.That(length, Is.EqualTo(3));
+
+			var readBuffer = new byte[6];
+			_ = await stream.ReadAsync(readBuffer.AsMemory(0, 6));
+			var newBuffer = new ushort[3];
+			Buffer.BlockCopy(readBuffer, 0, newBuffer, 0, 6);
+			Assert.That(newBuffer, Is.EquivalentTo(buffer));
+		}
+	}
+
+	[Test]
+	public static async Task WriteInt32ArrayAsync()
+	{
+		int[] buffer = [1, 2, 3];
+
+		var stream = new MemoryStream();
+		using var writer = new BinaryWriter(stream);
+		writer.Write((buffer.Length, buffer));
+
+		stream.Position = 0;
+
+		using (Assert.EnterMultipleScope())
+		{
+			var lengthBuffer = new byte[4];
+			_ = await stream.ReadAsync(lengthBuffer.AsMemory(0, 4));
+			var length = BitConverter.ToInt32(lengthBuffer);
+
+			Assert.That(length, Is.EqualTo(3));
+
+			var readBuffer = new byte[12];
+			_ = await stream.ReadAsync(readBuffer.AsMemory(0, 12));
+			var newBuffer = new int[3];
+			Buffer.BlockCopy(readBuffer, 0, newBuffer, 0, 12);
+			Assert.That(newBuffer, Is.EquivalentTo(buffer));
+		}
+	}
+
+	[Test]
+	public static async Task WriteUInt32ArrayAsync()
+	{
+		uint[] buffer = [1, 2, 3];
+
+		var stream = new MemoryStream();
+		using var writer = new BinaryWriter(stream);
+		writer.Write((buffer.Length, buffer));
+
+		stream.Position = 0;
+
+		using (Assert.EnterMultipleScope())
+		{
+			var lengthBuffer = new byte[4];
+			_ = await stream.ReadAsync(lengthBuffer.AsMemory(0, 4));
+			var length = BitConverter.ToInt32(lengthBuffer);
+
+			Assert.That(length, Is.EqualTo(3));
+
+			var readBuffer = new byte[12];
+			_ = await stream.ReadAsync(readBuffer.AsMemory(0, 12));
+			var newBuffer = new uint[3];
+			Buffer.BlockCopy(readBuffer, 0, newBuffer, 0, 12);
+			Assert.That(newBuffer, Is.EquivalentTo(buffer));
+		}
+	}
+
+	[Test]
+	public static async Task WriteInt64ArrayAsync()
+	{
+		long[] buffer = [1, 2, 3];
+
+		var stream = new MemoryStream();
+		using var writer = new BinaryWriter(stream);
+		writer.Write((buffer.Length, buffer));
+
+		stream.Position = 0;
+
+		using (Assert.EnterMultipleScope())
+		{
+			var lengthBuffer = new byte[4];
+			_ = await stream.ReadAsync(lengthBuffer.AsMemory(0, 4));
+			var length = BitConverter.ToInt32(lengthBuffer);
+
+			Assert.That(length, Is.EqualTo(3));
+
+			var readBuffer = new byte[24];
+			_ = await stream.ReadAsync(readBuffer.AsMemory(0, 24));
+			var newBuffer = new long[3];
+			Buffer.BlockCopy(readBuffer, 0, newBuffer, 0, 24);
+			Assert.That(newBuffer, Is.EquivalentTo(buffer));
+		}
+	}
+
+	[Test]
+	public static async Task WriteUInt64ArrayAsync()
+	{
+		ulong[] buffer = [1, 2, 3];
+
+		var stream = new MemoryStream();
+		using var writer = new BinaryWriter(stream);
+		writer.Write((buffer.Length, buffer));
+
+		stream.Position = 0;
+
+		using (Assert.EnterMultipleScope())
+		{
+			var lengthBuffer = new byte[4];
+			_ = await stream.ReadAsync(lengthBuffer.AsMemory(0, 4));
+			var length = BitConverter.ToInt32(lengthBuffer);
+
+			Assert.That(length, Is.EqualTo(3));
+
+			var readBuffer = new byte[24];
+			_ = await stream.ReadAsync(readBuffer.AsMemory(0, 24));
+			var newBuffer = new ulong[3];
+			Buffer.BlockCopy(readBuffer, 0, newBuffer, 0, 24);
+			Assert.That(newBuffer, Is.EquivalentTo(buffer));
 		}
 	}
 
@@ -93,7 +255,7 @@ internal static class BinaryWriterExtensionsTests
 		stream.Position = 0;
 
 		var lengthBuffer = new byte[8];
-	  _ = await stream.ReadAsync(lengthBuffer.AsMemory(0, 8));
+		_ = await stream.ReadAsync(lengthBuffer.AsMemory(0, 8));
 		var ticks = BitConverter.ToInt64(lengthBuffer);
 
 		Assert.That(ticks, Is.EqualTo(value.Ticks));
@@ -113,13 +275,13 @@ internal static class BinaryWriterExtensionsTests
 		using (Assert.EnterMultipleScope())
 		{
 			var ticksBuffer = new byte[8];
-		 _ = await stream.ReadAsync(ticksBuffer.AsMemory(0, 8));
+			_ = await stream.ReadAsync(ticksBuffer.AsMemory(0, 8));
 			var ticks = BitConverter.ToInt64(ticksBuffer);
 
 			Assert.That(ticks, Is.EqualTo(value.Ticks));
 
 			var offsetTicksBuffer = new byte[8];
-		 _ = await stream.ReadAsync(offsetTicksBuffer.AsMemory(0, 8));
+			_ = await stream.ReadAsync(offsetTicksBuffer.AsMemory(0, 8));
 			var offsetTicks = BitConverter.ToInt64(offsetTicksBuffer);
 
 			Assert.That(offsetTicks, Is.EqualTo(value.Offset.Ticks));
@@ -141,7 +303,7 @@ internal static class BinaryWriterExtensionsTests
 		using (Assert.EnterMultipleScope())
 		{
 			var buffer = new byte[valueBits.Length * 4];
-		 _ = await stream.ReadAsync(buffer.AsMemory(0, buffer.Length));
+			_ = await stream.ReadAsync(buffer.AsMemory(0, buffer.Length));
 
 			for (var i = 0; i < valueBits.Length; i++)
 			{
@@ -163,7 +325,7 @@ internal static class BinaryWriterExtensionsTests
 		stream.Position = 0;
 
 		var buffer = new byte[valueBuffer.Length];
-	  _ = await stream.ReadAsync(buffer.AsMemory(0, buffer.Length));
+		_ = await stream.ReadAsync(buffer.AsMemory(0, buffer.Length));
 
 		Assert.That(buffer, Is.EquivalentTo(valueBuffer));
 	}
@@ -180,7 +342,7 @@ internal static class BinaryWriterExtensionsTests
 		stream.Position = 0;
 
 		var ticksBuffer = new byte[8];
-	  _ = await stream.ReadAsync(ticksBuffer.AsMemory(0, 8));
+		_ = await stream.ReadAsync(ticksBuffer.AsMemory(0, 8));
 		var ticks = BitConverter.ToInt64(ticksBuffer);
 
 		Assert.That(ticks, Is.EqualTo(value.Ticks));